requirements.txt
numpy==1.23.2
opencv-python==4.5.2.54
opencv-contrib-python==4.5.2.54
PyQt6-Qt6==6.3.1